+++ packages/terra-framework-docs/src/terra-dev-site/doc/table/Examples/DefaultTable.jsx
@@ -53,6 +53,7 @@ const DefaultTable = () => (
     id="my-table-id"
     overflowColumns={tableData.cols}
     rows={tableData.rows}
+    ariaLabel="Table"
   />
 );