+++ packages/terra-table/src/subcomponents/ColumnHeader.jsx
@@ -24,9 +24,21 @@ const propTypes = {
   headerHeight: PropTypes.string.isRequired,
 
   /**
-   * Number that specifies the height of the table in pixels.
-   */
+  * Number that specifies the height of the table in pixels.
+  */
   tableHeight: PropTypes.number,
+  /**
+   * Column index for cell that can receive tab focus.
+   */
+  activeColumnIndex: PropTypes.number,
+  /**
+   * Specifies if resize handle should be active.
+   */
+  isActiveColumnResizing: PropTypes.bool,
+  /**
+   * Numeric increment in pixels to adjust column width when resizing via the keyboard.
+   */
+  columnResizeIncrement: PropTypes.number,
 
   /**
    * Function that is called when a selectable header cell is selected. Parameters:
@@ -38,6 +50,11 @@ const propTypes = {
    * Function that is called when the mouse down event is triggered on the column resize handle.
    */
   onResizeMouseDown: PropTypes.func,
+
+  /**
+   * Function that is called when the the keyboard is used to adjust the column size.
+   */
+  onResizeHandleChange: PropTypes.func,
 
   /**
    * Boolean indicating whether or not the table columns should be displayed.
@@ -52,11 +69,15 @@ const defaultProps = {
 const ColumnHeader = (props) => {
   const {
     tableId,
+    activeColumnIndex,
+    isActiveColumnResizing,
+    columnResizeIncrement,
     columns,
     headerHeight,
     tableHeight,
     onColumnSelect,
     onResizeMouseDown,
+    onResizeHandleChange,
     hasColumnHeaders,
   } = props;
 
@@ -81,10 +102,14 @@ const ColumnHeader = (props) => {
             isResizable={hasColumnHeaders && column.isResizable}
             isSelectable={hasColumnHeaders && column.isSelectable}
             tableHeight={tableHeight}
+            isActive={activeColumnIndex === columnIndex}
+            isResizeActive={activeColumnIndex === columnIndex && isActiveColumnResizing}
+            columnResizeIncrement={columnResizeIncrement}
             hasError={column.hasError}
             sortIndicator={column.sortIndicator}
             onColumnSelect={onColumnSelect}
             onResizeMouseDown={onResizeMouseDown}
+            onResizeHandleChange={onResizeHandleChange}
           />
         ))}
       </tr>