+++ packages/terra-framework-docs/CHANGELOG.md
@@ -9,7 +9,9 @@
   * Added examples and test to `terra-table` for row selection mode.
   * Added initial setup for `terra-compact-interactive-list` documentation.
   * Added an error handling test for `terra-embedded-content-consumer`.
-    
+  * Added single and ranged cell selection Flowsheet Data Grid example.
+  * Added aria labels for table examples.
+
 * Updated
   * Updated `terra-table` example to show providing custom masked cell message label.